+++ crates/large-smt-backend-rocksdb/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    miden_node_rocksdb_cxx_linkage_fix::configure();
-}